merge with upstream

+++ AnkiDroid/src/main/java/com/ichi2/libanki/Sched.java
@@ -1982,7 +1982,7 @@ public class Sched {
      * @return A string like “1 min” or “1.7 mo”
      */
     public String nextIvlStr(Context context, Card card, int ease) {
-        int ivl = nextIvl(card, ease);
+        long ivl = nextIvl(card, ease);
         if (ivl == 0) {
             return context.getString(R.string.sched_end);
         }
@@ -2001,7 +2001,7 @@ public class Sched {
     /**
      * Return the next interval for CARD, in seconds.
      */
-    public int nextIvl(Card card, int ease) {
+    public long nextIvl(Card card, int ease) {
         try {
             if (card.getQueue() == 0 || card.getQueue() == 1 || card.getQueue() == 3) {
                 return _nextLrnIvl(card, ease);
@@ -2009,12 +2009,12 @@ public class Sched {
                 // lapsed
                 JSONObject conf = _lapseConf(card);
                 if (conf.getJSONArray("delays").length() > 0) {
-                    return (int) (conf.getJSONArray("delays").getDouble(0) * 60.0);
+                    return (long) (conf.getJSONArray("delays").getDouble(0) * 60.0);
                 }
-                return _nextLapseIvl(card, conf) * 86400;
+                return _nextLapseIvl(card, conf) * 86400L;
             } else {
                 // review
-                return _nextRevIvl(card, ease) * 86400;
+                return _nextRevIvl(card, ease) * 86400L;
             }
         } catch (JSONException e) {
             throw new RuntimeException(e);
@@ -2022,7 +2022,7 @@ public class Sched {
     }
 
 
-    private int _nextLrnIvl(Card card, int ease) {
+    private long _nextLrnIvl(Card card, int ease) {
         // this isn't easily extracted from the learn code
         if (card.getQueue() == 0) {
             card.setLeft(_startingLeft(card));
@@ -2037,7 +2037,7 @@ public class Sched {
                 if (!_resched(card)) {
                     return 0;
                 }
-                return _graduatingIvl(card, conf, true, false) * 86400;
+                return _graduatingIvl(card, conf, true, false) * 86400L;
             } else {
                 int left = card.getLeft() % 1000 - 1;
                 if (left <= 0) {
@@ -2045,7 +2045,7 @@ public class Sched {
                     if (!_resched(card)) {
                         return 0;
                     }
-                    return _graduatingIvl(card, conf, false, false) * 86400;
+                    return _graduatingIvl(card, conf, false, false) * 86400L;
                 } else {
                     return _delayForGrade(conf, left);
                 }
